pyproject.toml
[tool.poetry]
name = "mywebsite"
version = "0.1.0"
description = ""
authors = ["Author Name <author_email@somedomain.com>"]

[tool.poetry.dependencies]
python = "^3.10"
ipython = "*"
pandas = "*"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
Markdown = "*"
django-filter = "^22.1"
djangorestframework-simplejwt = "*"
environs = "*"
gunicorn = "*"
importlib-metadata = "*"
psycopg2-binary = "*"
pymongo = "3.12.1"
djongo = "^1.3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
